src/multi-agents/modules/multiagents/requirements.txt
Flask==3.0.0
pyautogen
openai
influxdb-client
pandas